pyproject.toml
[project]
name = "tau-supersymmetry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "atlas-mpl-style>=0.25.0",
    "awkward>=2.9.0",
    "hydra-core>=1.3.2",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "matplotlib>=3.10.8",
    "mlflow>=3.9.0",
    "numpy>=2.4.2",
    "optuna>=4.7.0",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "pyrootutils>=1.0.4",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "uproot>=5.7.1",
    "xgboost>=3.2.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]

[tool.ruff]
extend-exclude = ["notebooks/*.ipynb"]
